+++ follower.hpp
@@ -5,14 +5,14 @@
 class follower : public position{
     private:
         car* connected_car;
-        float lenght;
-        float beta(float car_direction);
+        double lenght;
+        double beta(double car_direction);
         point last_hitch_pos;
     public:
-        float angle;
-        follower(car* bbc, float len, float hbeta);
+        double angle;
+        follower(car* bbc, double len, double hbeta);
         virtual  ~follower();
-        void set_to_car(float hbeta);
-        float beta();
+        void set_to_car(double hbeta);
+        double beta();
         void move();
 };
